requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
mysql-connector-python==8.0.18
protobuf==3.11.0
python-dotenv==0.10.3
requests==2.22.0
six==1.13.0
urllib3==1.25.7
